+++ app/comma-chameleon/views/index.html
@@ -31,6 +31,8 @@
       </div>
     </div>
     <script src="../hot.js"></script>
+    <script src="../ragged-rows.js"></script>
+    <script src="../validate.js"></script>
     <script src="../menu.js"></script>
   </body>
 </html>
